backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
httpx==0.27.2
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic-settings==2.5.2
numpy<2
faiss-cpu==1.8.0
sentence-transformers==3.1.1
PyPDF2==3.0.1
python-docx==1.1.2
